requirements.txt
quart
langchain
langchain_openai
python-dotenv